Rework into separate NullableInstanceOfRule level 8 rule
